+++ packages/app/src/systems/Core/components/AmountVisibility/AmountVisibility.tsx
@@ -3,13 +3,15 @@ import type { BigNumberish, BNInput } from 'fuels';
 import { formatAmount } from '../../utils';
 
 export type AmountVisibilityProps = {
-  value: BigNumberish | BNInput | undefined;
-  visibility: boolean | undefined;
+  value?: BigNumberish | BNInput;
+  units?: number;
+  visibility?: boolean;
 };
 
 export function AmountVisibility({
   visibility = true,
   value,
+  units,
 }: AmountVisibilityProps) {
-  return <>{visibility ? formatAmount(value) : '•••••'}</>;
+  return <>{visibility ? formatAmount(value, units) : '•••••'}</>;
 }