+++ packages/e2e-contract-tests/src/App.tsx
@@ -1,0 +1,34 @@
+import {
+  ForwardEthCard,
+  Header,
+  MintAssetCard,
+  Providers,
+  DepositHalfEthCard,
+  ForwardCustomAssetCard,
+  ForwardHalfAndMintCard,
+  ForwardHalfAndExternalMintCard,
+  ForwardHalfCustomAssetCard,
+  DepositAndMintMultiCalls,
+  AssetConfigurationCard,
+} from './components';
+
+function App() {
+  return (
+    <Providers>
+      <Header />
+      <div>
+        <MintAssetCard />
+        <ForwardEthCard />
+        <DepositHalfEthCard />
+        <ForwardCustomAssetCard />
+        <ForwardHalfAndExternalMintCard />
+        <ForwardHalfCustomAssetCard />
+        <ForwardHalfAndMintCard />
+        <DepositAndMintMultiCalls />
+        <AssetConfigurationCard />
+      </div>
+    </Providers>
+  );
+}
+
+export default App;