+++ packages/e2e-contract-tests/src/utils/assetId.ts
@@ -1,0 +1,12 @@
+import { BaseAssetId, Address, hash } from 'fuels';
+
+export const calculateAssetId = (
+  contractId: string,
+  subId: string = BaseAssetId
+) => {
+  const contractIdBytes = Address.fromAddressOrString(contractId).toBytes();
+  const subIdBytes = Address.fromAddressOrString(subId).toBytes();
+  const bytesToHash = Uint8Array.from([...contractIdBytes, ...subIdBytes]);
+  const assetId = hash(bytesToHash);
+  return assetId;
+};