+++ packages/react/src/hooks/useConnectors.tsx
@@ -1,14 +1,13 @@
 import { useQuery } from '@tanstack/react-query';
 
+import { useFuel } from '../components';
 import { QUERY_KEYS } from '../utils';
 
-import { useWindowFuel } from './useWindowFuel';
-
 export const useConnectors = () => {
-  const fuel = useWindowFuel();
+  const { fuel } = useFuel();
 
   const query = useQuery(
-    [QUERY_KEYS.chain, fuel],
+    [QUERY_KEYS.connectorList],
     async () => {
       return fuel?.listConnectors();
     },