+++ packages/types/src/asset.ts
@@ -5,6 +5,7 @@ export type Asset = {
   assetId: string;
   imageUrl?: string;
   symbol?: string;
+  decimals?: number;
   isCustom?: boolean;
 };